refactor: Split TestComponentSerializer into JSON and binary serializers
Extract TestJsonSerializer and TestBinarySerializer from the monolithic
TestComponentSerializer class (260 lines) to separate JSON and binary
serialization concerns.

- Create TestJsonSerializer implementing IComponentSerializer
- Create TestBinarySerializer implementing IBinaryComponentSerializer
- Refactor TestComponentSerializer to use composition, delegating to
  the specialized serializers while maintaining backward compatibility

Closes #482
